# openapi.yaml
openapi: 3.0.3
info:
  title: Dog Trainers Directory API
  description: |
    API for the Dog Trainers Directory platform.
    
    This API provides endpoints for:
    - Public access to trainer listings and search
    - Trainer profile management
    - Admin moderation and management
    - Emergency triage and contacts
    - Authentication with MFA support
    
    ## Authentication
    
    Most endpoints require authentication via JWT tokens. Include the token in the Authorization header:
    
    `Authorization: Bearer <access_token>`
    
    ## Rate Limiting
    
    All endpoints are rate limited. Rate limit headers are included in responses:
    
    - `X-RateLimit-Limit`: Maximum requests per window
    - `X-RateLimit-Remaining`: Remaining requests in current window
    - `X-RateLimit-Reset`: Unix timestamp when window resets
    - `Retry-After`: Seconds until retry (only when rate limited)
  version: 1.0.0
  contact:
    name: Dog Trainers Directory Support
    email: support@dogtrainersdirectory.com.au
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.dogtrainersdirectory.com.au/v1
    description: Production server
  - url: https://staging-api.dogtrainersdirectory.com.au/v1
    description: Staging server
  - url: http://localhost:3000/api/v1
    description: Development server
...
paths:
...
  /trainers:
    get:
      tags:
        - Public
      summary: Search trainers
      description: Search and filter trainers with pagination
      operationId: searchTrainers
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/SortBy'
        - $ref: '#/components/parameters/SortOrder'
        - $ref: '#/components/parameters/CouncilId'
        - $ref: '#/components/parameters/LocalityId'
        - $ref: '#/components/parameters/ResourceType'
        - $ref: '#/components/parameters/AgeSpecialty'
        - $ref: '#/components/parameters/BehaviorIssue'
        - $ref: '#/components/parameters/ServiceType'
        - $ref: '#/components/parameters/Verified'
        - $ref: '#/components/parameters/Claimed'
        - $ref: '#/components/parameters/Search'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TrainersResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '429':
          $ref: '#/components/responses/RateLimitExceeded'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /trainers/{id}/reviews:
    get:
      tags:
        - Public
      summary: Get trainer reviews
      description: Get all reviews for a specific trainer
      operationId: getTrainerReviews
      parameters:
        - $ref: '#/components/parameters/TrainerId'
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/SortBy'
        - $ref: '#/components/parameters/SortOrder'
        - $ref: '#/components/parameters/MinRating'
        - $ref: '#/components/parameters/MaxRating'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TrainerReviewsResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /trainer/services/{serviceType}:
    delete:
      tags:
        - Trainers
      summary: Remove service
      description: Remove a service type from trainer profile
      operationId: removeService
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ServiceType'
      responses:
        '200':
          description: Service removed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RemoveServiceResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token authentication
  parameters:
    Page:
      name: page
      in: query
      description: Page number for pagination
      required: false
      schema:
        type: integer
        minimum: 1
        default: 1
    Limit:
      name: limit
      in: query
      description: Number of items per page
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    SortBy:
      name: sort_by
      in: query
      description: Field to sort by
      required: false
      schema:
        type: string
        enum: [name, created_at, rating, verified]
    SortOrder:
      name: sort_order
      in: query
      description: Sort order
      required: false
      schema:
        type: string
        enum: [asc, desc]
        default: asc
    CouncilId:
      name: council_id
      in: query
      description: Filter by council ID
      required: false
      schema:
        type: string
        format: uuid
    LocalityId:
      name: locality_id
      in: query
      description: Filter by locality ID
      required: false
      schema:
        type: string
        format: uuid
    ResourceType:
      name: resource_type
      in: query
      description: Filter by resource type
      required: false
      schema:
        type: string
        enum: [trainer, behaviour_consultant, emergency_vet, urgent_care, emergency_shelter]
    AgeSpecialty:
      name: age_specialty
      in: query
      description: Filter by age specialty
      required: false
      schema:
        type: string
        enum:
          - Puppy (0–6 months)
          - Adolescent (6–18 months)
          - Adult (1.5–7 years)
          - Senior (7–10 years)
          - Any age
    BehaviorIssue:
      name: behavior_issue
      in: query
      description: Filter by behavior issue
      required: false
      schema:
        type: string
        enum:
          - Pulling on lead
          - Separation anxiety
          - Barking
          - Aggression
          - Jumping up on people
          - Recall issues
          - Socialisation
          - Chewing
          - Digging
          - House training
          - Fear/phobias
          - Other
    ServiceType:
      name: service_type
      in: query
      description: Filter by service type
      required: false
      schema:
        type: string
        enum:
          - Puppy training
          - Obedience training
          - Behaviour consultations
          - Group classes
          - Private training
    Verified:
      name: verified
      in: query
      description: Filter by verified status
      required: false
      schema:
        type: boolean
    Claimed:
      name: claimed
      in: query
      description: Filter by claimed status
      required: false
      schema:
        type: boolean
    Search:
      name: search
      in: query
      description: Search term
      required: false
      schema:
        type: string
        minLength: 1
        maxLength: 200
...
    MinRating:
      name: min_rating
      in: query
      description: Minimum rating
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 5
    MaxRating:
      name: max_rating
      in: query
      description: Maximum rating
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 5
    TrainerId:
      name: id
      in: path
      description: Trainer ID
      required: true
      schema:
        type: string
        format: uuid
...
  schemas:
...
    ResponseMeta:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
        request_id:
          type: string
        version:
          type: string
    ErrorResponse:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
        details:
          type: object
        timestamp:
          type: string
          format: date-time
    TrainersResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            trainers:
              type: array
              items:
                $ref: '#/components/schemas/Trainer'
            total:
              type: integer
            page:
              type: integer
            limit:
              type: integer
            has_more:
              type: boolean
        meta:
          $ref: '#/components/schemas/ResponseMeta'
    Trainer:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        resource_type:
          type: string
          enum: [trainer, behaviour_consultant, emergency_vet, urgent_care, emergency_shelter]
        locality_id:
          type: string
          format: uuid
        council_id:
          type: string
          format: uuid
        phone:
          type: string
        email:
          type: string
          nullable: true
        website:
          type: string
          format: uri
          nullable: true
        description:
          type: string
          nullable: true
        age_specialties:
          type: array
          items:
            type: string
        behavior_issues:
          type: array
          items:
            type: string
        service_type_primary:
          type: string
          nullable: true
        service_type_secondary:
          type: array
          items:
            type: string
          nullable: true
        abr_abn:
          type: string
          nullable: true
        emergency_phone:
          type: string
          nullable: true
        emergency_hours:
          type: string
          nullable: true
        emergency_services:
          type: array
          items:
            type: string
          nullable: true
        verified:
          type: boolean
        claimed:
          type: boolean
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    Review:
      type: object
      properties:
        id:
          type: string
          format: uuid
        business_id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
          nullable: true
        rating:
          type: integer
          minimum: 1
          maximum: 5
        comment:
          type: string
        moderation_status:
          type: string
          enum: [pending, approved, rejected]
        approved_at:
          type: string
          format: date-time
          nullable: true
        rejected_at:
          type: string
          format: date-time
          nullable: true
        rejection_reason:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    TrainerReviewsResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            reviews:
              type: array
              items:
                $ref: '#/components/schemas/Review'
            total:
              type: integer
            page:
              type: integer
            limit:
              type: integer
            has_more:
              type: boolean
            trainer_id:
              type: string
              format: uuid
            average_rating:
              type: number
            total_reviews:
              type: integer
        meta:
          $ref: '#/components/schemas/ResponseMeta'
...
    RemoveServiceResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            trainer:
              $ref: '#/components/schemas/Trainer'
            removed_service:
              type: string
        meta:
          $ref: '#/components/schemas/ResponseMeta'
...
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    RateLimitExceeded:
      description: Rate limit exceeded
      headers:
        X-RateLimit-Limit:
          schema:
            type: integer
        X-RateLimit-Remaining:
          schema:
            type: integer
        X-RateLimit-Reset:
          schema:
            type: integer
        Retry-After:
          schema:
            type: integer
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'